requirements.txt
psycopg2-binary==2.8.5
python-logstash==0.4.6
